Normalized config vs netrc vs utils
