+++ lib/keys.js
@@ -7,7 +7,7 @@ var path = require('path')
     ;
 module.exports = {
     getDefaultKey: function() {
-        var sshPath = path.join(utils.getUserHome(), '.ssh')
+        var sshPath = path.join(configtools.getUserHome(), '.ssh')
             , rsaPath = path.join(sshPath, 'id_rsa.pub')
             , dsaPath = path.join(sshPath, 'id_dsa.pub')
             , hasRSA = fs.existsSync(rsaPath)
@@ -25,55 +25,49 @@ module.exports = {
         return fs.readFileSync(path, 'utf8');
     },
     sendKey: function(key, callback) {
-        configtools.getConfig(function(err, config) {    
-            var baseUrl = configtools.getBaseUrl(config)
-                , sendUrl = url.resolve(baseUrl, 'auth/add-key')
-                ;
-            if(process.env.DEXTER_DEV) {
-                console.log('Sending key to', sendUrl);
-            }
-            rest.post(sendUrl, {
-                data: { key: key }
-                , headers: configtools.getAuthorizedRequestHeaders(config)
-            }).on('complete', function(result, response) {
-                utils.restResponse(result, response, function() {
-                    if(callback) callback(result);
-                });
+        var baseUrl = configtools.getBaseUrl()
+            , sendUrl = url.resolve(baseUrl, 'auth/add-key')
+            ;
+        if(process.env.DEXTER_DEV) {
+            console.log('Sending key to', sendUrl);
+        }
+        rest.post(sendUrl, {
+            data: { key: key }
+            , headers: configtools.getAuthorizedRequestHeaders()
+        }).on('complete', function(result, response) {
+            utils.restResponse(result, response, function() {
+                if(callback) callback(result);
             });
         });
     },
     removeKey: function(query, callback) {
-        configtools.getConfig(function(err, config) {    
-            var baseUrl = configtools.getBaseUrl(config)
-                , sendUrl = url.resolve(baseUrl, 'auth/remove-key')
-                ;
-            if(process.env.DEXTER_DEV) {
-                console.log('Removing key:', sendUrl);
-            }
-            rest.del(sendUrl, {
-                data: { query: query }
-                , headers: configtools.getAuthorizedRequestHeaders(config)
-            }).on('complete', function(result, response) {
-                utils.restResponse(result, response, function() {
-                    if(callback) callback(result);
-                });
+        var baseUrl = configtools.getBaseUrl()
+            , sendUrl = url.resolve(baseUrl, 'auth/remove-key')
+            ;
+        if(process.env.DEXTER_DEV) {
+            console.log('Removing key:', sendUrl);
+        }
+        rest.del(sendUrl, {
+            data: { query: query }
+            , headers: configtools.getAuthorizedRequestHeaders()
+        }).on('complete', function(result, response) {
+            utils.restResponse(result, response, function() {
+                if(callback) callback(result);
             });
         });
     },
     getAll: function(callback) {
-        configtools.getConfig(function(err, config) {    
-            var baseUrl = configtools.getBaseUrl(config)
-                , sendUrl = url.resolve(baseUrl, 'auth/list-keys')
-                ;
-            if(process.env.DEXTER_DEV) {
-                console.log('Fetching keys:', sendUrl);
-            }
-            rest.get(sendUrl, {
-                headers: configtools.getAuthorizedRequestHeaders(config)
-            }).on('complete', function(result, response) {
-                utils.restResponse(result, response, function() {
-                    if(callback) callback(result);
-                });
+        var baseUrl = configtools.getBaseUrl()
+            , sendUrl = url.resolve(baseUrl, 'auth/list-keys')
+            ;
+        if(process.env.DEXTER_DEV) {
+            console.log('Fetching keys:', sendUrl);
+        }
+        rest.get(sendUrl, {
+            headers: configtools.getAuthorizedRequestHeaders()
+        }).on('complete', function(result, response) {
+            utils.restResponse(result, response, function() {
+                if(callback) callback(result);
             });
         });
     }
