+++ lib/netrc.js
@@ -4,6 +4,7 @@ var netrcreader = require('netrc')
     , url = require('url')
     , lazy = require('lazy')
     , EOL = require('os').EOL
+    , configtools = require('./config')
     , homepath
     , filename
     ;
@@ -16,15 +17,9 @@ filename = join(homepath, '.netrc');
 netrc = netrcreader(filename);
 
 module.exports = {
-    machineName: process.env.DEXTER_HOST || 'rundexter.com'
-    , hostIsHTTPS: process.env.DEXTER_HOST_HTTPS == 1
-    , urlRoot: function() {
-        var protocol = (this.hostIsHTTPS) ? 'https://' : 'http://';
-        return url.resolve(protocol + this.machineName, '/api/');
-    } 
     , getPassword: function() {
         try {
-            return netrc['this.machineName'].password;
+            return netrc[configtools.machineName].password;
         } catch(e) {
             if(process.env.DEXTER_DEV) {
                 console.log('Failed to extract netrc password:', e);
@@ -38,7 +33,6 @@ module.exports = {
             , oldContents = ''
             , newFile = fs.openSync(tmpName, 'w', 0600)
             , skipping = false
-            , _this = this
             ;
         if(fs.existsSync(filename)) {
             oldContents = fs.readFileSync(filename, { encoding: 'utf8' });
@@ -50,13 +44,13 @@ module.exports = {
             if(line.match(/^\s*machine /)) {
                 if(skipping) skipping = false;
                 //We're in our territory if the line ends with our name
-                skipping = line.match(new RegExp(_this.machineName + '\s*$'));
+                skipping = line.match(new RegExp(configtools.machineName + '\s*$'));
             }
             if(!skipping) {
                 fs.writeSync(newFile, line + EOL);
             }
         });
-        fs.writeSync(newFile, 'machine ' + this.machineName + EOL);
+        fs.writeSync(newFile, 'machine ' + configtools.machineName + EOL);
         fs.writeSync(newFile, '  login ' + login + EOL);
         fs.writeSync(newFile, '  password ' + password + EOL);
         fs.closeSync(newFile);
