+++ lib/utils.js
@@ -32,9 +32,6 @@ module.exports = {
             callback(null, config);
         });
     }
-    , getUserHome: function() {
-        return process.env.HOME || process.env.HOMEPATH || process.env.USERPROFILE;
-    }
     , streamToString: function(cb) {
         var content = '';
         process.stdin.resume();
